set up webpack wnvironment; set up jest library
