--- assets/game-multiplayer.css
@@ -1,225 +1,0 @@
-.pulse {
-  --color: #ef6eae;
-  --hover: #ef8f6e; }
-
-#game-multiplayer {
-  background: #483e6a;
-  background: linear-gradient(180deg, #1f0575 40%, #c34b6c 88%);
-  height: 100%;
-  overflow: hidden; }
-  #game-multiplayer nav {
-    display: flex;
-    justify-content: space-between;
-    z-index: 5;
-    position: fixed;
-    top: 0px;
-    width: 100%;
-    background: #160746;
-    height: 4em; }
-    #game-multiplayer nav button {
-      background: none;
-      border: none;
-      padding: 0 30px; }
-    #game-multiplayer nav h1 {
-      display: flex;
-      justify-content: center;
-      color: white;
-      font-size: 25px;
-      font-weight: bold;
-      background: #3d2c75;
-      border-radius: 22px;
-      height: 75%;
-      margin: 8px;
-      text-align: center;
-      padding-top: 4px;
-      width: 20%; }
-      @media screen and (max-width: 420px) {
-        #game-multiplayer nav h1 {
-          width: 8em; } }
-      #game-multiplayer nav h1 img {
-        width: 15px;
-        width: 40px;
-        margin-right: 10px; }
-  #game-multiplayer .navbar-img {
-    width: 30px; }
-  #game-multiplayer #message-box {
-    background-color: #3d2c75;
-    z-index: 5;
-    position: fixed;
-    top: 6em;
-    left: 34%;
-    display: flex;
-    width: 38em;
-    height: 3em;
-    border-radius: 5px;
-    border-top: 5px solid #160746;
-    padding: 10px 20px;
-    color: white; }
-  #game-multiplayer #orange-planet {
-    top: 12em;
-    right: 34%;
-    width: 803px;
-    position: relative;
-    filter: invert(100%) sepia(0%);
-    border-radius: 50%; }
-    @media screen and (max-width: 1000px) {
-      #game-multiplayer #orange-planet {
-        width: 28em;
-        top: 19em;
-        right: 44%; } }
-    @media screen and (max-width: 420px) {
-      #game-multiplayer #orange-planet {
-        width: 17em;
-        top: 16em;
-        right: 58%; } }
-  #game-multiplayer #purple-planet {
-    top: -24em;
-    left: 75%;
-    width: 353px;
-    position: relative; }
-    @media screen and (max-width: 1000px) {
-      #game-multiplayer #purple-planet {
-        width: 17em;
-        top: -2em;
-        right: 58%; } }
-    @media screen and (max-width: 420px) {
-      #game-multiplayer #purple-planet {
-        width: 8em;
-        top: 0em;
-        right: 0%; } }
-  #game-multiplayer #aqua-planet {
-    bottom: -31em;
-    left: 22%;
-    width: 653px;
-    position: relative; }
-    @media screen and (max-width: 1000px) {
-      #game-multiplayer #aqua-planet {
-        width: 23em;
-        top: 6em;
-        left: 42em; } }
-    @media screen and (max-width: 420px) {
-      #game-multiplayer #aqua-planet {
-        width: 14em;
-        top: 8em;
-        left: 63%; } }
-  #game-multiplayer .game-container {
-    z-index: 5;
-    position: fixed;
-    top: 11em;
-    left: 38%;
-    display: flex;
-    width: 30em;
-    justify-content: space-around;
-    text-align: center;
-    color: white;
-    font-weight: bold; }
-    @media screen and (max-width: 1000px) {
-      #game-multiplayer .game-container {
-        width: 28em;
-        top: 7em;
-        left: 27%; } }
-    @media screen and (max-width: 420px) {
-      #game-multiplayer .game-container {
-        width: 18em;
-        top: 7em;
-        left: 18%; } }
-    #game-multiplayer .game-container .profile-img {
-      border-radius: 50%;
-      background-color: white;
-      margin: 8px 0 10px 0; }
-      @media only screen and (max-width: 420px) {
-        #game-multiplayer .game-container .profile-img {
-          width: 6em; } }
-    #game-multiplayer .game-container .player-0,
-    #game-multiplayer .game-container .player-1 {
-      width: 120px; }
-    #game-multiplayer .game-container .player-0-img {
-      border: 10px solid #e9488b; }
-    #game-multiplayer .game-container .player-1-img {
-      border: 10px solid #59deff; }
-    #game-multiplayer .game-container .turn-marker {
-      width: 28px; }
-  #game-multiplayer .board {
-    width: 35em;
-    height: 22em;
-    transform: perspective(22px) rotateX(1deg);
-    margin: 50px;
-    position: fixed;
-    top: 40%;
-    left: 34%;
-    border-radius: 17px;
-    border-bottom: 10px solid black; }
-    @media only screen and (max-width: 1000px) {
-      #game-multiplayer .board {
-        width: 30em;
-        height: 22em;
-        left: 20%;
-        top: 35%; } }
-    @media only screen and (max-width: 420px) {
-      #game-multiplayer .board {
-        width: 15em;
-        height: 13em;
-        left: 13%; } }
-    #game-multiplayer .board .grid {
-      display: flex;
-      flex-wrap: wrap;
-      width: 100%;
-      height: 100%;
-      height: 100%; }
-      #game-multiplayer .board .grid div {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        font-size: 4em;
-        font-weight: bold;
-        background-color: #848AC4;
-        width: 33.3%;
-        height: 33.3%; }
-      #game-multiplayer .board .grid #grid-1,
-      #game-multiplayer .board .grid #grid-3,
-      #game-multiplayer .board .grid #grid-5,
-      #game-multiplayer .board .grid #grid-7,
-      #game-multiplayer .board .grid #grid-9 {
-        background-color: white; }
-      #game-multiplayer .board .grid #grid-0 {
-        border-radius: 17px 0 0 0; }
-      #game-multiplayer .board .grid #grid-2 {
-        border-radius: 0 17px 0 0; }
-      #game-multiplayer .board .grid #grid-6 {
-        border-radius: 0 0 0 7px; }
-      #game-multiplayer .board .grid #grid-8 {
-        border-radius: 0 0 7px 0; }
-  #game-multiplayer .score-0 {
-    background-color: #e9488b;
-    position: fixed;
-    bottom: 0;
-    left: 33%; }
-    @media only screen and (max-width: 1000px) {
-      #game-multiplayer .score-0 {
-        left: 24%; } }
-    @media only screen and (max-width: 420px) {
-      #game-multiplayer .score-0 {
-        left: 17%; } }
-  #game-multiplayer .score-1 {
-    background-color: #59deff;
-    position: fixed;
-    bottom: 0;
-    left: 59%; }
-  #game-multiplayer .score-0,
-  #game-multiplayer .score-1 {
-    height: 42px;
-    padding: 0 19px;
-    border-radius: 20px 20px 0 0;
-    color: white;
-    font-weight: bold;
-    display: flex;
-    align-items: center; }
-    @media only screen and (max-width: 420px) {
-      #game-multiplayer .score-0,
-      #game-multiplayer .score-1 {
-        width: 8em; } }
-  #game-multiplayer .trophy-img {
-    width: 20px;
-    margin-right: 8px; }
-
-/*# sourceMappingURL=game-multiplayer.css.map */
