--- assets/game-multiplayer.scss
@@ -1,251 +1,0 @@
-@import 'variables';
-
-#game-multiplayer {
-    background: rgb(72,62,106);
-    background: linear-gradient(180deg, $nightblue 40%, $hardpink 88%);
-    height: 100%;
-    overflow: hidden;
-    nav {
-        display: flex;
-        justify-content: space-between;
-        z-index: 5;
-        position: fixed;
-        top: 0px;
-        width: 100%;
-        background: $navblue;
-        height: 4em;
-        button {
-            background: none;
-            border: none;
-            padding: 0 30px;
-        }
-        h1 { 
-            display: flex;
-            justify-content: center;
-            color: white;
-            font-size: 25px;
-            font-weight: bold;
-            background: #3d2c75;
-            border-radius: 22px;
-            height: 75%;
-            margin: 8px;
-            text-align: center;
-            padding-top: 4px;
-            width: 20%;
-            @media screen and (max-width:420px) { 
-                width: 8em;
-                
-            }
-            img {
-                width: 15px;
-                width: 40px;
-                margin-right: 10px;
-            }
-        }
-    }
-    .navbar-img {
-        width: 30px;
-    }
-    #message-box {
-        background-color: #3d2c75;
-        z-index: 5;
-        position: fixed;
-        top: 6em;
-        left: 34%;
-        display: flex;
-        width: 38em;
-        height: 3em;
-        border-radius: 5px;
-        border-top: 5px solid $navblue;
-        padding: 10px 20px;
-        color: white;
-    }
-    #orange-planet {
-        top: 12em;
-        right: 34%;
-        width: 803px;
-        position: relative;
-        filter: invert(100%) sepia(0%);
-        border-radius: 50%; 
-        @media screen and (max-width:1000px) { 
-            width: 28em;
-            top: 19em;
-            right: 44%;
-        }
-        @media screen and (max-width:420px) { 
-            width: 17em;
-            top: 16em;
-            right: 58%;
-        }
-    }
-    #purple-planet {
-        top: -24em;
-        left: 75%;
-        width: 353px;
-        position: relative;
-        @media screen and (max-width:1000px) { 
-            width: 17em;
-            top: -2em;
-            right: 58%;
-        }
-        @media screen and (max-width:420px) { 
-            width: 8em;
-            top: 0em;
-            right: 0%; 
-        }
-    }
-    #aqua-planet {
-        bottom: -31em;
-        left: 22%;
-        width: 653px;
-        position: relative;
-        @media screen and (max-width:1000px) { 
-            width: 23em;
-            top: 6em;
-            left: 42em;
-        }
-        @media screen and (max-width:420px) { 
-            width: 14em;
-            top: 8em;
-            left: 63%;
-        }  
-    }
-    .game-container {
-        z-index: 5;
-        position: fixed;
-        top: 11em;
-        left: 38%;
-        @media screen and (max-width:1000px) { 
-            width: 28em;
-            top: 7em;
-            left: 27%
-        }
-        @media screen and (max-width:420px) { 
-            width: 18em;
-            top: 7em;
-            left: 18%;
-        }
-        
-        display: flex;
-        width: 30em;
-        justify-content: space-around;
-        text-align: center;
-        color: white;
-        font-weight: bold;
-        .profile-img {
-            border-radius: 50%;
-            background-color: white;
-            margin: 8px 0 10px 0;
-            @media only screen and (max-width:420px) { 
-                width: 6em; 
-            }
-        }
-        .player-0,
-        .player-1 {
-            width: 120px;
-        }
-        .player-0-img {
-            border: 10px solid $pink;
-        }
-        .player-1-img {
-            border: 10px solid $blue;
-        }
-        .turn-marker {
-            width: 28px;
-        }
-    }
-    .board {
-        width: 35em;
-        height: 22em;
-        transform: perspective(22px) rotateX(1deg);
-        margin: 50px;
-        position: fixed;
-        top: 40%;
-        left: 34%;
-        border-radius: 17px;
-        border-bottom: 10px solid black;
-        @media only screen and (max-width:1000px) { 
-            width: 30em;
-            height: 22em;
-            left: 20%;
-            top: 35%;
-        }
-        @media only screen and (max-width:420px) { 
-            width: 15em;
-            height: 13em;
-            left: 13%;
-        }
-        .grid {
-            display: flex;
-            flex-wrap: wrap;
-            width: 100%;
-            height: 100%;
-            height: 100%;
-            div {
-                display: flex;
-                justify-content: center;
-                align-items: center;
-                font-size: 4em;
-                font-weight: bold;	                            
-                background-color: $purple;	
-                width: 33.3%;
-                height: 33.3%;
-            }	            
-            #grid-1,	
-            #grid-3,	
-            #grid-5,	
-            #grid-7,	
-            #grid-9 {	
-                background-color: white;	
-            }	
-            #grid-0 {	
-                border-radius: 17px 0 0 0;	
-            }	
-            #grid-2 {	
-                border-radius: 0 17px 0 0;	
-            }	
-            #grid-6 {	
-                border-radius: 0 0 0 7px;	
-            }	
-            #grid-8 {	
-                border-radius: 0 0 7px 0;	
-            }	
-        }
-        
-    }
-    .score-0 {
-        background-color: #e9488b;
-        position: fixed;
-        bottom: 0;
-        left: 33%;
-        @media only screen and (max-width:1000px) { 
-            left: 24%; 
-        }
-        @media only screen and (max-width:420px) { 
-            left: 17%; 
-        }
-    }
-    .score-1 {
-        background-color: $blue;
-        position: fixed;
-        bottom: 0;
-        left: 59%;
-    }
-    .score-0,
-    .score-1 {
-        height: 42px;
-        padding: 0 19px;
-        border-radius: 20px 20px 0 0; 
-        color: white;
-        font-weight: bold;   
-        display: flex;
-        align-items: center;
-        @media only screen and (max-width:420px) { 
-            width: 8em; 
-        }
-    }
-    .trophy-img {
-        width: 20px;
-        margin-right: 8px;
-    }
-}
