--- assets/variables.css
@@ -1,5 +1,0 @@
-.pulse {
-  --color: #ef6eae;
-  --hover: #ef8f6e; }
-
-/*# sourceMappingURL=variables.css.map */
