--- index.html
@@ -1,99 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="assets/global.css">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css"/>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bulma/0.5.1/css/bulma.min.css"/>
-    
-    <!-- 
-    <script src="./js/lib/p5.js" type="text/javascript"></script>
-    <script src="./js/lib/p5.play.js" type="text/javascript"></script>  
-    <script src="./js/sketch.js" type="text/javascript"></script>
-    <script src="./js/bubble.js" type="text/javascript"></script>
-
-    -->
-    
-    <title>Tic-tac-toe</title>
-</head>
-<body>
-    <main id="menu">
-        <div class="new-game">
-            <div id="logo">
-                <div><span id="pink-color">X</span>O</div>
-                <div>T<span id="blue-color">O</span></div>
-            </div>
-            <div id="menu-btns">
-                <button class="pulse" id="multiplayer-btn">
-                    <div id="logo">
-                        <div><span id="pink-color">X</span>O</div>
-                        <div>T<span id="blue-color">O</span></div>
-                    </div>
-                    <p>MULTIPLAYER</p>
-                </button>
-                <div class="hidden" id="form">
-                    <div class="field">
-                        <label class="pink-color">Player 1:</label>
-                        <input class="input" name="name-0" type="text" placeholder="name">
-                    </div>
-                    <div class="field">
-                        <label class="blue-color">Player 2:</label>
-                        <input class="input" name="name-1" type="text" placeholder="name">
-                    </div>
-                </div>
-                    
-                <button class="pulse" id="youvsai-btn">
-                    <img src="./assets/img/brain.svg" alt="brain-img">
-                    <p>YOU VS AI</p>
-                </button>
-                <button class="pulse" id="rules-btn">
-                    <img src="./assets/img/document.svg" alt="document-img">
-                    <p>RULES</p>
-                </button>
-        
-            </div>    
-        </div>    
-    </main>
-    <main id="game-ai">
-
-    </main>
-    <main id="game-multiplayer" class="hidden">
-        <img src="./assets/img/neptune-purple.svg" id="purple-planet">
-        <img src="./assets/img/neptune.svg" id="orange-planet">
-        <img src="./assets/img/neptune.svg" id="aqua-planet">
-        <nav>
-            <button id="back"><img class="navbar-img" src="./assets/img/back.svg"></button>
-            <h1><img src="./assets/img/tic-tac-toe.svg">Tic-Tac-Toe</h1>
-            <button id="replay"><img class="navbar-img" src="./assets/img/refresh.svg"></button>
-        </nav>
-        <div id="message-box" class="hidden">
-            <h1>Message</h1>
-        </div>
-        <div class="game-container">
-            <div class="player-0">
-                <div class="name-0" id="name-0"></div>
-                <img src="./assets/img/brain.svg" alt="" class="player-0-img profile-img">
-                <div class="turn-0"><img class="turn-marker" src="./assets/img/arrow-pink.svg"></div>
-                <div class="turn-ball-0 hidden">#</div> 
-            </div>
-            <div class="player-1">
-                <div class="name-1" id="name-1"></div>
-                <img src="./assets/img/brain.svg" alt="" class="player-1-img profile-img">
-                <div class="turn-1 hidden"><img class="turn-marker" src="./assets/img/arrow-blue.svg"></div>
-                <div class="turn-ball-1 hidden">#</div>
-            </div>
-            
-        </div>
-        <div class="board">
-            <div class="grid" id="grid">
-                
-            </div>
-        </div>        
-        <div class="score-0"><img class="trophy-img" src="./assets/img/trophy.svg"><p id="score-0">0 won matches</p> </div>
-        <div class="score-1"><img class="trophy-img" src="./assets/img/trophy.svg"><p id="score-1">0 won matches</p> </div>
-    </main>
-</body>
-
-<script src="js/gameModule.js" type="module"></script>
-</html>
