+++ src/assets/global.scss
@@ -1,0 +1,31 @@
+@import 'game-multiplayer';
+@import 'menu';
+@import 'variables';
+
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+html {
+    height: 100%;
+    overflow-y:hidden;
+}
+
+body {
+    height: 100%;
+    overflow-y:hidden;
+}
+
+.pink-color {
+    color: $pink;
+}
+
+.blue-color {
+    color: $blue;
+}
+
+.hidden {
+    display: none !important;
+}
