+++ src/assets/menu.scss
@@ -1,0 +1,115 @@
+@import 'variables';
+
+#menu {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    background: rgb(72,62,106);
+    background: linear-gradient(180deg, $nightblue 40%, $hardpink 88%);
+    .new-game {
+        display: flex;
+        align-items: center;
+        flex-direction: column;
+        #logo {
+            z-index: 3;
+            font-size: 100px;
+            font-weight: bolder;
+            line-height: 80px;
+            color: white;
+            #pink-color {
+                color: $pink;
+            }
+            #blue-color {
+                color: $blue;
+            }
+        }
+        #menu-btns {
+            margin-top: 4em;
+            z-index: 3;
+            button {
+                padding: 0 1em;
+                display: flex;
+                justify-content: baseline;
+                align-items: center;
+                width: 70em;
+                
+                @media only screen 
+                    and (min-device-width: 320px) 
+                    and (max-device-width: 480px)
+                    and (-webkit-min-device-pixel-ratio: 2) {
+                       height: 5em; 
+                       width: 22em
+                    }
+                height: 6em;
+                border-radius: 18px;
+                border: none;
+                margin-bottom: 1.5em;
+                transition: 0.25s;
+                -webkit-box-shadow: 0px 44px 68px -12px rgba(0,0,0,0.75);
+                -moz-box-shadow: 0px 44px 68px -12px rgba(0,0,0,0.75);
+                box-shadow: 0px 44px 68px -12px rgba(0,0,0,0.75);
+                p {
+                    color: white;
+                    font-size: 21px;
+                    font-weight: bold;
+                    margin-left: 20px;
+                }
+                img {
+                    width: 50px;
+                    filter: invert(100%) sepia(0%);
+                }
+                &:hover, 
+                &:focus {
+                    cursor: pointer;
+                    animation: pulse 1s;
+                    box-shadow: 0 0 0 2em rgba(#fff,0);
+                    border-color: var(--hover);
+                    color: #fff;
+                    
+                    
+                }
+            }
+            @keyframes pulse {
+                0% { box-shadow: 0 0 0 0 var(--hover); }
+            }
+
+            #multiplayer-btn {
+                background-color: $blue;
+                #logo {
+                    font-size: 21px;
+                    line-height: 23px;
+                    margin: 10px;
+
+                    #pink-color {
+                        color: white;
+                    }
+                    #blue-color {
+                        color: white;
+                    }
+                }
+            }
+            #youvsai-btn {
+                background-color: $pink;
+
+            } 
+            #rules-btn {
+                background-color: $yellow;
+            }
+        }
+    }
+    #form {
+        display: flex;
+        justify-content: space-evenly; 
+        margin-bottom: 15px;
+        .field {
+            flex-direction: column;
+            text-align: center;
+            label {
+                font-size: 20px;
+                font-weight: bold;
+            }
+        }
+    }
+}
