+++ src/js/players.js
@@ -1,0 +1,6 @@
+const playerFactory = (name, mark, score) => {
+  score = 0;
+  return { name, mark, score };
+};
+
+export default playerFactory;
